backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
pydantic-settings==2.10.1
sqlalchemy==2.0.30
asyncpg==0.29.0
pymongo
bcrypt==4.2.1
PyJWT==2.10.1
openai>=1.0.0
celery==5.4.0
redis==5.0.7
httpx==0.27.0
groq==1.1.2
requests
langgraph==0.2.24
sentence-transformers==3.0.1
pandas
numpy
scikit-learn
joblib
networkx
prophet
faker
faiss-cpu
pypdf
pillow
pytesseract
pdf2image
python-multipart==0.0.9